pyproject.toml
[project]
name = "storia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pratham Das", email = "prathamdas003@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "groq>=1.1.2",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[project.scripts]
storia= "storia.main:main"

[tool.setuptools.package-data]
"storia.laguages" = ["*.json"]
